chore(team): split remote relay route helpers
